feat: support runtime filter push down to Table AM
when seq scan begins, check whether the scanflags of table am is set to
determine whether the runtime filter is pushed down.

When the runtime filter is pushed down to pax am, pax am converts the min/max
scankey in the runtime filter into PFTNode and performs min/max filtering.