PAX: Rename all PAX guc names from pax_xxx to pax.xxx
The original guc names can't be set in configuration file, because
the configuration file is loaded before the PAX gucs are registered.

The gucs are usually used in session level. But setting the guc
in configuration file will cause critical issue, so we rename
the guc name to pax.xxx. The name containing the dot('.') is considered
as a placeholder if the guc is not found.